add logtime in panic recover and makes some log to debug
