+++ pkg/log/errorlog.go
@@ -18,16 +18,8 @@
 package log
 
 import (
-	"strconv"
-	"sync/atomic"
-	"time"
-
 	"sofastack.io/sofa-mosn/pkg/types"
-)
-
-var (
-	// lastTime is used to cache time
-	lastTime atomic.Value
+	"sofastack.io/sofa-mosn/pkg/utils"
 )
 
 // errorLogger is a default implementation of ErrorLogger
@@ -150,30 +142,7 @@ func (l *errorLogger) GetLogLevel() Level {
 	return l.level
 }
 
-// timeCache is used to reduce format
-type timeCache struct {
-	t int64
-	s string
-}
-
-// We use a cache to reduce the format
+// logTime is a wrapper of utils.CacheTime for compatibility
 func logTime() string {
-	var s string
-	t := time.Now()
-	nano := t.UnixNano()
-	now := nano / 1e9
-	value := lastTime.Load()
-	if value != nil {
-		last := value.(*timeCache)
-		if now <= last.t {
-			s = last.s
-		}
-	}
-	if s == "" {
-		s = t.Format("2006-01-02 15:04:05")
-		lastTime.Store(&timeCache{now, s})
-	}
-	mi := nano % 1e9 / 1e6
-	s = s + "," + strconv.Itoa(int(mi))
-	return s
+	return utils.CacheTime()
 }
