+++ pkg/proxy/upstream.go
@@ -268,8 +268,8 @@ func (r *upstreamRequest) OnFailure(reason types.PoolFailureReason, host types.H
 
 func (r *upstreamRequest) OnReady(sender types.StreamSender, host types.Host) {
 	// debug message for upstream
-	if log.Proxy.GetLogLevel() >= log.INFO {
-		log.Proxy.Infof(r.downStream.context, "[proxy] [upstream] connPool ready, proxyId = %v, host = %s", r.downStream.ID, host.AddressString())
+	if log.Proxy.GetLogLevel() >= log.DEBUG {
+		log.Proxy.Debugf(r.downStream.context, "[proxy] [upstream] connPool ready, proxyId = %v, host = %s", r.downStream.ID, host.AddressString())
 	}
 
 	r.requestSender = sender
